requirements.txt
autogluon.common==1.4.0
autogluon.core==1.4.0
autogluon.features==1.4.0
autogluon.tabular==1.4.0
boto3==1.42.24
botocore==1.42.24
catboost==1.2.8
category-encoders==2.6.4
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.1
graphviz==0.21
idna==3.11
importlib_resources==6.5.2
impyute==0.0.8
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.4
lightgbm==4.6.0
matplotlib==3.9.4
narwhals==2.15.0
networkx==3.2.1
numpy==1.25.2
packaging==23.1
pandas==2.3.3
patsy==1.0.2
Pillow==10.0.0
plotly==6.5.1
psutil==7.0.0
pyarrow==20.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.32.5
s3transfer==0.16.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.16.0
statsmodels==0.14.6
threadpoolctl==3.2.0
torch==1.8.1
tqdm==4.64.1
typing_extensions==4.7.1
tzdata==2025.3
urllib3==1.26.20
xgboost==2.1.4
zipp==3.23.0